% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cycle_length.R
\name{calculate_cycle_length}
\alias{calculate_cycle_length}
\title{Calculate the shortest cycle length}
\usage{
calculate_cycle_length(graph, x, y, connected = FALSE)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{x}{A node.}

\item{y}{A node.}
}
\description{
This function calculate the length of a shortest cycle between two nodes. If these nodes
don't form a cycle, the function returns length 0.
}
\details{
The function has dependency on igraph.
}
\keyword{cycle}
\keyword{shortest}
