% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combn_ind.R
\name{combn_ind}
\alias{combn_ind}
\title{Auxilliary function to wrs_bktr()}
\usage{
combn_ind(i, n)
}
\arguments{
\item{i}{Index of the number of visits. i > 1.}

\item{n}{Total number of edges could be safely removed. n > 1.}
}
\description{
Auxilliary function to wrs_bktr()
}
