% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_nbrs.R
\name{common_nbrs}
\alias{common_nbrs}
\title{A function that finds the common neighbours between two nodes}
\usage{
common_nbrs(graph, x, y)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{x}{A node in graph.}

\item{y}{A node in graph that is different from x.}
}
\description{
This function finds the common neighbours between two nodes from a graph. It
returns an empty vector of length 0 if there is no common neighbours.
}
