% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_wrs_deg4.R
\name{d_wrs_deg4}
\alias{d_wrs_deg4}
\title{Definite-WRS for maximum degree 4 graphs}
\usage{
d_wrs_deg4(G, debug = F)
}
\arguments{
\item{G}{A given undirected graph stored in an adjacency matrix format.}

\item{debug}{A boolean argument to show steps of the backtracking algorithm.}
}
\description{
This function is a polynomial time algorithm for checking the morality of
undirected graphs with maximum degree 4. The algorithm has been proved to be
correct. It either removes a simplicial node or a simplicial clique, depending
on situations like node degree, and K3 stack length etc.
}
