% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deficiency.R
\name{deficiency}
\alias{deficiency}
\title{A function to calculates the deficiency of a node}
\usage{
deficiency(graph, x)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{node}{A node that is expressed in text, e.g. "V1" or "X1".}
}
\description{
The deficiency of a node in a graph is the number of edges, adding which make
the node's neighbours a complete subgraph (or a clique).
}
\keyword{deficiency}
