% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{A function to calculates the degree of a node}
\usage{
degree(graph, node)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{node}{A node that is expressed in text, e.g. "V1" or "X1".}
}
\description{
A function to calculates the degree of a node
}
\keyword{degree}
