% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_dist_graph.R
\name{edit_dist_graph}
\alias{edit_dist_graph}
\title{The edit distance from the learned to the true undirected graph
Both graphs are stored in adjacency matrix. The outcome is a list consists
of tp, fp, tn, fn, precision, recall, F1 score and edit distance.}
\usage{
edit_dist_graph(learned, true, debug = FALSE)
}
\arguments{
\item{learned}{The learned graph stored in an adjacency matrix format.}

\item{true}{The true graph stored in an adjacency matrix format.}

\item{debug}{A boolean argument to show steps of the backtracking algorithm.}
}
\description{
The edit distance from the learned to the true undirected graph
Both graphs are stored in adjacency matrix. The outcome is a list consists
of tp, fp, tn, fn, precision, recall, F1 score and edit distance.
}
