% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nbr.R
\name{find_nbr}
\alias{find_nbr}
\title{A function that finds neighbours of a node}
\usage{
find_nbr(graph, node)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{node}{A node that is expressed in text, e.g. "V1" or "X1".}
}
\description{
This function finds neighbours of a node in a graph, which is stored in a contigency matrix. The output
of the function are a vector of node names.
}
\keyword{neighbour}
