% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_simplicial.R
\name{find_simplicial}
\alias{find_simplicial}
\title{Find simplicial nodes}
\usage{
find_simplicial(graph)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}
}
\description{
This function finds all simplicial nodes in a graph. Notice that isolated nodes are treated as
simplicial nodes for programming convenience. The output is a vector of node names, ordered by
node degree ascending.
}
\keyword{simplicial}
