% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_ordering_moralization.R
\name{fixed_ordering_moralization}
\alias{fixed_ordering_moralization}
\title{Fixed ordering moralization}
\usage{
fixed_ordering_moralization(graph, ordering)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}
}
\description{
This function turns a immoral graph to a moral graph by fill in the deficiency
of each node for a given node ordering. The elimination process removes a
nodes and the edges between its neighbours.
The output of the function is a moral graph.
}
\keyword{moralization}
