% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_chordal.R
\name{is_chordal}
\alias{is_chordal}
\title{Test chordality}
\usage{
is_chordal(graph)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}
}
\description{
This function tests if a given graph is a chordal graph. A chordal graph is an undirected graph
such that any cycle of length greater than 3 must have a chord. It has dependency on the function
is_chordal in the igraph package.
}
\keyword{chordal}
