% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_clique.R
\name{is_clique}
\alias{is_clique}
\title{Identify a clique}
\usage{
is_clique(graph, nodes)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{nodes}{A set of nodes that are expressed in texts, e.g. c("V1", "V2")}
}
\description{
This function tests if a set of nodes form a clique. A clique is a complete (fully connected) graph. The
output of the function is a boolean variable with a value 0 or 1.
}
\keyword{clique}
