% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_simplicial.R
\name{is_simplicial}
\alias{is_simplicial}
\title{Identify a simplicial node}
\usage{
is_simplicial(graph, node)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{node}{A node that is expressed in text, e.g. "V1" or "X1".}
}
\description{
This function tests if a given node is a simplicial node in a graph. A simplicial node is a node, whose
neighbours form a clique. The output of the function is a boolean variable with a value 0 or 1.
}
\keyword{simplicial}
