% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k3_ladder.R
\name{k3_ladder}
\alias{k3_ladder}
\title{A function that look for a k3 ladder.}
\usage{
k3_ladder(graph, s, x, y, l)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{s}{A simplicial node in the graph.}

\item{x}{One neighbour of the node s.}

\item{y}{The other neighbour of the node s.}

\item{l}{The number of k3s adjacent togethers. It starts from 1.}
}
\description{
This function finds the length of the k3 (triangle) ladder starts from the
given simplicial node. It returns the number of adjancet k3s. It also returns
the value -1 if the initial k3 shares an edge with either a k4 or two k3s,
and value 0 if the initial k3 and 3 other k3s form a triangle.
}
