% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_special_envelop_graph.R
\name{matching_special_envelop_graph}
\alias{matching_special_envelop_graph}
\title{special subgraph matching}
\usage{
matching_special_envelop_graph(G, node)
}
\arguments{
\item{G}{A given undirected graph stored in an adjacency matrix format.}

\item{node}{A node that is in a stack of 3 K3s.}
}
\description{
This function searches for a special envelop subgraph in the given graph G.
The function returns 0 if there is not such a matching or 1 if there is or NA
if the given node is not in a stack of 3 K3s.
}
