% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mblist2moral.R
\name{mblist2moral}
\alias{mblist2moral}
\title{Convert a list of MBs to an undirected graph
The list of MBs must be symmetric in order for the converted graph to be
symmetric.}
\usage{
mblist2moral(mbList, vars)
}
\arguments{
\item{mbList}{A list of MBs.}

\item{vars}{A vector of all the variables. Must be in the same order as the
list name.}
}
\description{
Convert a list of MBs to an undirected graph
The list of MBs must be symmetric in order for the converted graph to be
symmetric.
}
