% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_deg_moralization.R
\name{min_deg_moralization}
\alias{min_deg_moralization}
\title{Minimum degree moralization}
\usage{
min_deg_moralization(graph)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}
}
\description{
This function turns a immoral graph to a moral graph by fill in the deficiency
of each node. A node with the minimum degree is selected at each step.
The elimination process removes a nodes and the edges between its neighbours.
The output of the function is a moral graph.
}
\keyword{moralization}
