% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_deg_triangulation.R
\name{min_deg_triangulation}
\alias{min_deg_triangulation}
\title{Minimum degree triangulation}
\usage{
min_deg_triangulation(graph)
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}
}
\description{
This function turns a graph to a triangulated graph by fill in the deficiency
of each node. A node with the minimum degree is selected at each step.
The elimination process removes a simplicial node at a time.
The output of the function is a triangulated graph.
}
\keyword{triangulation}
