% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moralization.R
\name{moralization}
\alias{moralization}
\title{Graph moralization}
\usage{
moralization(graph, ordering, excess = c(0, 1))
}
\arguments{
\item{graph}{A graph that is stored in an adjacency matrix.}

\item{ordering}{A node ordering.}

\item{excess}{The edges (between the neighbours of the simplicial node) that
are removed with the simplicial node. At the moment, this is a binary input
that takes 0 or 1. The value 0 corresponds to no edge removal. The resulting
graph is triangulated. The value 1 corresponds to remove all edges between
the neighbours of each simplicial node.}
}
\description{
This function turns an immoral graph to a moral graph by following the node
ordering given. It either removes no or all edges between the neighbours of
the simplicial node at each step of moraliziing the graph.
}
\keyword{moralization}
