% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{A function to keep track of the progress in a loop}
\usage{
progress(i, q, n)
}
\arguments{
\item{i}{Index in a loop.}

\item{q}{Separation, e.g., only show every 100 indices.}

\item{n}{Total number of iterations.}
}
\description{
The function keeps track of the looping progress.
}
