% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_bounded_deg_graph.R
\name{rand_bounded_deg_graph}
\alias{rand_bounded_deg_graph}
\title{Random graph generation}
\usage{
rand_bounded_deg_graph(n, delta)
}
\arguments{
\item{n}{The number of nodes in the graph.}

\item{delta}{The maximum degree in the graph.}
}
\description{
This function uses the rand_bounded_deg_graph_aux() to generate a random undirected graph for a
given maximum degree. The maximum degree may not be reached. This function uses try() to catch
error when there is no consistent DAG for a sampled degree sequence. There is always an output
from this function.
}
\keyword{bounded}
\keyword{degree}
\keyword{graph,}
\keyword{random}
