% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_bounded_deg_graph_aux.R
\name{rand_bounded_deg_graph_aux}
\alias{rand_bounded_deg_graph_aux}
\title{Random graph generation (auxiliary)}
\usage{
rand_bounded_deg_graph_aux(n, delta)
}
\arguments{
\item{n}{The number of nodes in the graph.}

\item{delta}{The maximum degree in the graph.}
}
\description{
This is an auxiliary function of the rand_bounded_deg_graph() function. It may return an error due to
non-existence of a consistent DAG for a sampled degree sequence. It first generates a random
degree sequence based on the maximum degree, then use the igraph::sample_degseq(out.deg, method = "vl")
function to generate a connected undirected random graph consistents with the sampled degree sequence.
}
\keyword{bounded}
\keyword{degree}
\keyword{graph,}
\keyword{random}
