% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicated_edges.R
\name{remove_duplicated_edges}
\alias{remove_duplicated_edges}
\title{A function to remove duplicated rows in a dataframe or matrix}
\usage{
remove_duplicated_edges(x)
}
\arguments{
\item{x}{A dataframe or row.}
}
\description{
This function removes from a dataframe or matrix duplicated rows. The dataframe or matrix stores
edges of a graph. The dataframe or matrix has two columns, which are the two end points of an edge.
}
