% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrs_bktr.R
\name{wrs_bktr}
\alias{wrs_bktr}
\title{Test weak recursively simplicial}
\usage{
wrs_bktr(G, D, debug = F)
}
\arguments{
\item{G}{A given undirected graph stored in an adjacency matrix format.}

\item{D}{An undirected graph to start orienting with. Coud be set to G.}

\item{debug}{A boolean argument to show steps of the backtracking algorithm.}
}
\description{
This function tests if a given graph is weak recursively simplicial. If test positive, the given
undirected graph is also oriented into a hybrid graph (a graph may contain both undirected and
directed edges). This function uses backtracking method to test a graph, so it can be slow for large
graphs. It outputs positive if the graph is either chordal or wrs.
}
\keyword{recursively}
\keyword{simplicial}
\keyword{weak}
