% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_wrs.R
\name{d_wrs}
\alias{d_wrs}
\title{Definite-WRS for maximum degree 3 graphs}
\usage{
d_wrs(G, debug = F)
}
\arguments{
\item{G}{A given undirected graph stored in an adjacency matrix format.}

\item{debug}{A boolean argument to show steps of the backtracking algorithm.}
}
\description{
This function is a polynomial time algorithm for checking the morality of
undirected graphs with maximum degree 3. The algorithm has been proved to be
correct. It finds a simplicial clique (i.e., a simplicial node
and its neighbours), and removes the entire clique.
}
